Cleanup and generalization; functions are moved to their own files to allow for generalization
