+++ lib/display/markdown-mathjax.js
