+++ lib/display/markdownRenderer.js
@@ -1,5 +1,5 @@
 import markdownIt from "markdown-it";
-import markdownItMathjax from "../lib/markdown-mathjax";
+import markdownItMathjax from "./markdown-mathjax";
 import { markdownItSub } from "markdown-it-sub";
 
 //
@@ -10,6 +10,7 @@ let cleanMarkdown = (text) => {
     /--------------------------------------------------------------------------------/g,
     ""
   ); // This is how *** gets translated to. Per se, it should be translated to "\n", but not doing so to keep consistent.
+  //text = text.replace(/\n/g, "\n\n")
   text = text.replace("... Read more »", "...");
   //text = text.replaceAll("\n", "\n\n") // Does not work; original text doesn't have newlines (?!). Search for Debate on Instrumental Convergence between LeCun, Russell, Bengio, Zador, and More and see https://www.alignmentforum.org/posts/WxW6Gc6f2z3mzmqKs/debate-on-instrumental-convergence-between-lecun-russell to confirm.
   return text;
