+++ lib/getProps/processAlignmentNewsletterBlurb.js
@@ -47,8 +47,8 @@ export function processAlignmentNewsletterBlurb({ blurb, papers }) {
 
 /* Testing
 // Test in the final app by searching for a paper with weird Rohin symbols, like "How can Interpretability help Alignment?"
-import { getPapers } from "../lib/airtablegraph.js";
-const { papers } = await getPapers();
+import { getData } from "../lib/getData.js";
+const { papers } = await getData();
 //console.log(papers)
 
 let blurb = `Interpretability seems to be useful for a wide variety of AI alignment proposals. Presumably, different proposals require different kinds of interpretability. This post analyzes this question to allow researchers to prioritize across different kinds of interpretability research.
