+++ pages/layout.js
@@ -2,6 +2,11 @@ import Link from "next/link";
 import Head from "next/head";
 import { GiPapers } from "react-icons/gi";
 
+const appName = "AI Safety Papers"
+const iconLocation = "/papers.png"
+const Icon = GiPapers
+const feedbackPage = "https://github.com/QURIresearch/ai-safety-papers/discussions"
+
 export default function Layout(props) {
   const classNameSelected = (isSelected) =>
     `text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-sm text-md font-medium cursor-pointer ${
@@ -10,8 +15,8 @@ export default function Layout(props) {
   return (
     <div>
       <Head>
-        <title>AI Safety Papers</title>
-        <link rel="icon" href="/papers.png" />
+        <title>{appName}</title>
+        <link rel="icon" href={iconLocation} />
       </Head>
 
       <div>
@@ -24,9 +29,9 @@ export default function Layout(props) {
                   <Link href={`/`} passHref>
                     <span className="flex items-center">
                       <span className="mr-2">
-                        <GiPapers />
+                        <Icon />
                       </span>
-                      AI Safety Papers
+                      {appName}
                     </span>
                   </Link>
                 </div>
@@ -66,7 +71,7 @@ export default function Layout(props) {
                 <div className="md:block">
                   <div className="ml-2 flex items-baseline space-x-4">
                     <a
-                      href="https://github.com/QURIresearch/ai-safety-papers/discussions"
+                      href={"https://github.com/QURIresearch/ai-safety-papers/discussions"}
                       className={classNameSelected(false)}
                       target="_blank"
                     >
